# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EHPS API",
    "description" : "Equipment Health & Performance System — semiconductor fab equipment maintenance. Covers fab-coordinator machine/technician management, routine checkups, alert escalation and approval, repair assignment/completion, and the equipment health dashboard. All endpoints (except /api/auth/**) require a bearer JWT obtained from POST /api/auth/login, and are scoped by role: FAB_COORDINATOR, MANAGER, or TECHNICIAN. See docs/API-CONTRACT.md for the full endpoint contract.",
    "version" : "2.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/fab/machines/{machineId}" : {
      "get" : {
        "tags" : [ "Fab — Machines" ],
        "summary" : "Get a machine by ID",
        "description" : "Returns a single machine owned by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "getMachineById",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Fab — Machines" ],
        "summary" : "Update a machine",
        "description" : "Updates an existing machine owned by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "updateMachine",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateMachineRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/technician/checkups/machines/{machineId}/results" : {
      "post" : {
        "tags" : [ "Technician — Checkups" ],
        "summary" : "Submit checkup results",
        "description" : "Records the results of a performed checkup on the given machine, raising alerts if parameters are out of range. Requires role TECHNICIAN.",
        "operationId" : "performCheckup",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PerformCheckupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePerformCheckupResult"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/fab/machines" : {
      "get" : {
        "tags" : [ "Fab — Machines" ],
        "summary" : "List my machines",
        "description" : "Returns all machines owned by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "getMyMachines",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Fab — Machines" ],
        "summary" : "Register a new machine",
        "description" : "Adds a new machine owned by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "addMachine",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddMachineRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/fab/checkup-assignments" : {
      "post" : {
        "tags" : [ "Fab — Checkup Assignments" ],
        "summary" : "Assign a technician to a checkup",
        "description" : "Assigns a technician to perform a routine checkup on a machine. Requires role FAB_COORDINATOR.",
        "operationId" : "assignTechnicianForCheckup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssignCheckupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCheckupAssignmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/technician/repairs/machines/{machineId}" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "Get machine details for a repair",
        "description" : "Returns machine details for a machine assigned to the calling technician for repair. Requires role TECHNICIAN.",
        "operationId" : "getMachineDetails",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineDetailsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/technician/checkups/machines/{machineId}" : {
      "get" : {
        "tags" : [ "Technician — Checkups" ],
        "summary" : "Get machine details for a checkup",
        "description" : "Returns machine details for a machine assigned to the calling technician for checkup. Requires role TECHNICIAN.",
        "operationId" : "getMachineDetails_1",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineDetailsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/technician/checkup-assignments" : {
      "get" : {
        "tags" : [ "Technician — Checkups" ],
        "summary" : "List my assigned checkup work",
        "description" : "Returns checkups currently assigned to the calling technician. Requires role TECHNICIAN.",
        "operationId" : "getAssignedCheckupWorks",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAssignedWorkResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/dashboard/equipment" : {
      "get" : {
        "tags" : [ "Dashboard" ],
        "summary" : "Get equipment history",
        "description" : "Returns equipment checkup/repair history within an optional date range and machine type filter. Requires role MANAGER or FAB_COORDINATOR.",
        "operationId" : "getDashboardHistory",
        "parameters" : [ {
          "name" : "fromDate",
          "in" : "query",
          "description" : "Start date (inclusive), ISO-8601 (yyyy-MM-dd)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "toDate",
          "in" : "query",
          "description" : "End date (inclusive), ISO-8601 (yyyy-MM-dd)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "typeId",
          "in" : "query",
          "description" : "Machine type ID filter; 0 means all types",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListDashboardHistoryResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "UpdateMachineRequest" : {
        "type" : "object",
        "properties" : {
          "machineCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "machineCode", "typeId" ]
      },
      "ApiResponseMachineResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/MachineResponse"
          }
        }
      },
      "MachineResponse" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "typeName" : {
            "type" : "string"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PerformCheckupRequest" : {
        "type" : "object",
        "properties" : {
          "values" : {
            "type" : "array",
            "items" : {
              "type" : "number",
              "format" : "float"
            }
          }
        },
        "required" : [ "values" ]
      },
      "ApiResponsePerformCheckupResult" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PerformCheckupResult"
          }
        }
      },
      "PerformCheckupResult" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineType" : {
            "type" : "string"
          },
          "finalHealth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "statuses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "riskAlertCreated" : {
            "type" : "boolean"
          },
          "riskAlertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "severity" : {
            "type" : "string"
          }
        }
      },
...
      "AddMachineRequest" : {
        "type" : "object",
        "properties" : {
          "machineCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "machineCode", "typeId" ]
      },
      "AssignCheckupRequest" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "machineId", "technicianId" ]
      },
      "ApiResponseCheckupAssignmentResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CheckupAssignmentResponse"
          }
        }
      },
      "CheckupAssignmentResponse" : {
        "type" : "object",
        "properties" : {
          "workId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workType" : {
            "type" : "string"
          },
          "workDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "ApiResponseMachineDetailsResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/MachineDetailsResponse"
          }
        }
      },
      "MachineDetailsResponse" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "typeName" : {
            "type" : "string"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ApiResponseListAssignedWorkResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssignedWorkResponse"
            }
          }
        }
      },
      "AssignedWorkResponse" : {
        "type" : "object",
        "properties" : {
          "workId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workType" : {
            "type" : "string"
          },
          "workDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "ApiResponseListMachineResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MachineResponse"
            }
          }
        }
      },
...
      "ApiResponseListDashboardHistoryResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DashboardHistoryResponse"
            }
          }
        }
      },
      "DashboardHistoryResponse" : {
        "type" : "object",
        "properties" : {
          "historyId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "machineType" : {
            "type" : "string"
          },
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "technicianName" : {
            "type" : "string"
          },
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "historyDate" : {
            "type" : "string",
            "format" : "date"
          },
          "issue" : {
            "type" : "string"
          },
          "repairAction" : {
            "type" : "string"
          },
          "observations" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}